+++ tests/testthat/test-runs-negbinomial.R
@@ -23,7 +23,7 @@ good_negbinomial = list(
     y ~ 1 + x + shape(1),
     ~ 0 + x + shape(1)
   ),
-  list(y ~ 1 + ar(1))
+  list(y ~ 1 + ar(1) + ma(1))
 )
 
 test_good(